jaja-code-node-structure/src/main/java/edu/ubfc/st/vm/project/grp7/jaja/code/ast/node/JcStopNode.java
package edu.ubfc.st.vm.project.grp7.jaja.code.ast.node;

import edu.ubfc.st.vm.project.grp7.ast.ASTNode;
import edu.ubfc.st.vm.project.grp7.ast.Breakpoint;
import edu.ubfc.st.vm.project.grp7.jaja.code.ast.JajaCodeNode;

public interface JcStopNode extends JajaCodeNode {

    static Builder builder() {
        return new Builder();
    }

    class Builder extends JajaCodeNode.NodeBuilder<JcStopNode.Builder> {
        public JcStopNode build() {
            return new JcStopImpl(this.line, this.column, this.breakpoint);
        }
    }

}
